requirements.txt
asgiref==3.8.1
beautifulsoup4==4.12.3
certifi==2024.8.30
chardet==5.2.0
charset-normalizer==3.3.2
crispy-bootstrap5==2024.2
Django==4.2.15
django-crispy-forms==2.3
django-environ==0.11.2
django-filter==24.3
django_csp==3.8
gunicorn==23.0.0
idna==3.10
itsdangerous==2.2.0
packaging==24.1
pillow==10.4.0
psycopg2-binary==2.9.10
reportlab==4.2.5
requests==2.32.3
soupsieve==2.6
sqlparse==0.5.1
typing_extensions==4.12.2
urllib3==2.2.3
whitenoise==6.7.0
